Refactor the application
